Flask/requirements.txt
flask
flask-MySQLdb
pymongo
flask_cors
flask_restful
pyodbc
pandas
pymssql
psycopg2